test: Suppress benign forkpty deprecation warning in pytest
The warning 'This process is multi-threaded, use of forkpty() may lead
to deadlocks' is a Python 3.13 addition that's mostly a false alarm for
our short-lived test processes.

As recommended by experts, we can safely suppress this warning for now
and plan migration to Shellous before Python 3.14 when it may become
an error.